the print all commit file

+++ 0x10-variadic_functions/3-print_all.c
@@ -33,7 +33,7 @@ void print_all(const char * const format, ...)
 					break;
 
 				case 'f';
-					printf("%s%f", sep, va_arg(list, int));
+					printf("%s%f", sep, va_arg(list, double));
 					break;
 
 				case 's';
